fix(vidpalette): byte-truncate add-affect multipliers in RealizeResource/3d
Recovered RealizeResource (0x7C04B0) and RealizeResource3d (0x7D6240)
computed range-add multipliers as ~GetRValue(c) on INT fields, where
int ~0 = -1. The binary truncates to a byte before the multiply:
(uint)(byte)~bVar * inv, i.e. 255 for an un-added channel. With -1,
the screen-blend combine forced red/blue to 0 and halved green, so
any additive color glow (elemental hit ColorGlowDissipate) rendered
the sprite as a dark-green/black silhouette instead of brightening it.

Both the period==1 branch and the pulse branch are fixed; the pulse
branch needs a double cast, (BYTE)~(BYTE)(pct*chan/period), matching
the binary which truncates the scaled channel before inverting.
RealizeRange/RealizeRange3d already truncated correctly via BYTE
RGBQUAD assignments.

Also align CPersistantEffectColorEffect::AIUpdate (0x4C15F0) mid-tick
condition with the binary: post-increment m_counter < m_period (the
recovery added +1 again, ending the fade one tick early).

Document sym.py scan <hexbytes> in the CLAUDE.md lookup table.

+++ src/CVidPalette.cpp
@@ -1144,18 +1144,18 @@ void CVidPalette::RealizeResource(DWORD* pDestPalette, INT nBpp, DWORD dwFlags, 
     if (pAffectArgs != NULL) {
         if (pAffectArgs->pRangeAdds[0] != NULL) {
             if (pAffectArgs->aRangeAddPeriods[0] == 1 || pAffectArgs->aRangeAddPeriods[0] == 0) {
-                rgbInv.rgbRed *= ~GetRValue(*pAffectArgs->pRangeAdds[0]);
-                rgbInv.rgbGreen *= ~GetGValue(*pAffectArgs->pRangeAdds[0]);
-                rgbInv.rgbBlue *= ~GetBValue(*pAffectArgs->pRangeAdds[0]);
+                rgbInv.rgbRed *= static_cast<BYTE>(~GetRValue(*pAffectArgs->pRangeAdds[0]));
+                rgbInv.rgbGreen *= static_cast<BYTE>(~GetGValue(*pAffectArgs->pRangeAdds[0]));
+                rgbInv.rgbBlue *= static_cast<BYTE>(~GetBValue(*pAffectArgs->pRangeAdds[0]));
             } else {
                 nAddPercentage = m_nAUCounter % (2 * pAffectArgs->aRangeAddPeriods[0]);
                 if (nAddPercentage > pAffectArgs->aRangeAddPeriods[0]) {
                     nAddPercentage = 2 * pAffectArgs->aRangeAddPeriods[0] - nAddPercentage;
                 }
 
-                rgbInv.rgbRed *= ~(nAddPercentage * GetRValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]);
-                rgbInv.rgbGreen *= ~(nAddPercentage * GetGValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]);
-                rgbInv.rgbBlue *= ~(nAddPercentage * GetBValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]);
+                rgbInv.rgbRed *= static_cast<BYTE>(~static_cast<BYTE>(nAddPercentage * GetRValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]));
+                rgbInv.rgbGreen *= static_cast<BYTE>(~static_cast<BYTE>(nAddPercentage * GetGValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]));
+                rgbInv.rgbBlue *= static_cast<BYTE>(~static_cast<BYTE>(nAddPercentage * GetBValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]));
             }
 
             bAdd = TRUE;
@@ -1377,18 +1377,18 @@ void CVidPalette::RealizeResource3d(DWORD* pDestPalette, DWORD dwFlags, CVIDIMG_
     if (pAffectArgs != NULL) {
         if (pAffectArgs->pRangeAdds[0] != NULL) {
             if (pAffectArgs->aRangeAddPeriods[0] == 1 || pAffectArgs->aRangeAddPeriods[0] == 0) {
-                rgbInv.rgbRed *= ~GetRValue(*pAffectArgs->pRangeAdds[0]);
-                rgbInv.rgbGreen *= ~GetGValue(*pAffectArgs->pRangeAdds[0]);
-                rgbInv.rgbBlue *= ~GetBValue(*pAffectArgs->pRangeAdds[0]);
+                rgbInv.rgbRed *= static_cast<BYTE>(~GetRValue(*pAffectArgs->pRangeAdds[0]));
+                rgbInv.rgbGreen *= static_cast<BYTE>(~GetGValue(*pAffectArgs->pRangeAdds[0]));
+                rgbInv.rgbBlue *= static_cast<BYTE>(~GetBValue(*pAffectArgs->pRangeAdds[0]));
             } else {
                 nAddPercentage = m_nAUCounter % (2 * pAffectArgs->aRangeAddPeriods[0]);
                 if (nAddPercentage > pAffectArgs->aRangeAddPeriods[0]) {
                     nAddPercentage = 2 * pAffectArgs->aRangeAddPeriods[0] - nAddPercentage;
                 }
 
-                rgbInv.rgbRed *= ~(nAddPercentage * GetRValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]);
-                rgbInv.rgbGreen *= ~(nAddPercentage * GetGValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]);
-                rgbInv.rgbBlue *= ~(nAddPercentage * GetBValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]);
+                rgbInv.rgbRed *= static_cast<BYTE>(~static_cast<BYTE>(nAddPercentage * GetRValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]));
+                rgbInv.rgbGreen *= static_cast<BYTE>(~static_cast<BYTE>(nAddPercentage * GetGValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]));
+                rgbInv.rgbBlue *= static_cast<BYTE>(~static_cast<BYTE>(nAddPercentage * GetBValue(*pAffectArgs->pRangeAdds[0]) / pAffectArgs->aRangeAddPeriods[0]));
             }
 
             bAdd = TRUE;
